# File: Re-add support to skip CR (`\r`) in `File::get_as_text`