--- a/.gitattributes
+++ b/.gitattributes
@@ -7,6 +7,8 @@ thirdparty/* linguist-vendored
 * text=auto eol=lf
 # Except for bat files, which are Windows only files
 *.bat eol=crlf
+# And some test files where the EOL matters
+*.test.txt -text
 
 # The above only works properly for Git 2.10+, so for older versions
 # we need to manually list the binary files we don't want modified.