--- a/tests/data/line_endings_cr.test.txt
+++ b/tests/data/line_endings_cr.test.txt
@@ -1,0 +1,1 @@
+Hello darknessMy old friendI've come to talkWith you again