--- a/tests/data/line_endings_lf.test.txt
+++ b/tests/data/line_endings_lf.test.txt
@@ -1,0 +1,4 @@
+Hello darkness
+My old friend
+I've come to talk
+With you again