This was removed in #63481, and we confirmed that it's better like this,
but we add back the possibility to strip CR as an option, to optionally
restore the previous behavior.

For performance this is done directly in `String::parse_utf8`.

Also fixes Android `FileAccess::get_line()` as this one _should_ strip CR.

Supersedes #63717.

Will backport for `3.x` with the `get_as_text()` option defaulting to true to fix the compat breakage.

@bruvzg My changes to `String::parse_utf8()` might be a huge hack, please review.

---

Speed comparison on a big file with a debug build on Linux, as described in https://github.com/godotengine/godot/pull/63717#issuecomment-1200440103:

- Original `File::get_as_text()` from before #63481

```
Unix file: 3.107707 s
DOS file:  3.214315 s
```

- Current `master` after #63481

```
Unix file: 0.136303 s
DOS file:  0.162586 s
```

- PR #63717 with `replace("\r", "")`

```
Unix file: 0.180824 s
DOS file:  1.624483 s
```

- This PR

```
Unix file: 0.129991 s
DOS file, keep CR: 0.148745 s
DOS file, skip CR: 0.190492 s
```