@akien-mga 2022-07-31
I still had a doubt about my `parse_utf8` changes so I added some tests, and indeed they fail :P