@bruvzg 2022-07-31
Test files seem to be wrong, CRLF one it only has LF.