@akien-mga 2022-07-31
> Test files seem to be wrong, CRLF one it only has LF.

Indeed, my `.gitattributes` override doesn't seem to work as intended. The Git documentation is horrible to read :|

```
./tests/core/string/test_string.h:155:
TEST CASE:  [String] UTF8 with CR

./tests/core/string/test_string.h:166: ERROR: CHECK( no_cr == base.replace("\r", "") ) is NOT correct!
  values: CHECK( H == Hello darkness
My old friend
I've come to talkWith you again )
```