@DarkKilauea 2022-07-31
Should an enumeration that selects the desired line ending be used instead?  I understand this is a quick fix, but I'm concerned about adding a parameter that will need to be changed later on, resulting in a breaking change down the line.

``` c
enum LineEnding {
    LineEnding_Unix, // lf
    LineEnding_Mac // cr
    LineEnding_Windows // crlf
}
```