@akien-mga 2022-07-31
Probably overkill, no modern platform uses CR only anymore. It was used on MacOS prior to MacOS X.
Since MacOS X (2001), it's using LF like other Unixes. Windows uses CRLF.